requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
cinemagoer @ git+https://github.com/cinemagoer/cinemagoer@c33292d692f139e17df76121d7de54dcd67324bb
click==8.1.7
click-default-group==1.2.4
dacite==1.8.1
diskcache==5.6.3
dunamai==1.19.2
filelock==3.13.1
frozenlist==1.3.3
idna==3.4
jsonpickle==2.2.0
langcodes==3.3.0
language-data==1.1
lxml==4.9.3
marisa-trie==0.7.8
multidict==6.0.4
packaging==23.2
parse-torrent-name==1.1.1
pathvalidate==3.2.0
pyd2v==1.3.0
PyYAML==6.0.1
requests==2.31.0
requests-file==2.0.0
SQLAlchemy==2.0.27
tldextract==3.5.0
tmdbsimple==2.9.1
typing_extensions==4.10.0
Unidecode==1.3.8
urllib3==2.0.3
xmltodict==0.13.0
yarl==1.9.2
